screens + usecases

+++ domain/src/main/java/com/ilyakrn/foodies/domain/usecases/RemoveProductFromBasketUseCase.kt
@@ -11,7 +11,15 @@ import com.ilyakrn.foodies.domain.repositories.TagRepository
 class RemoveProductFromBasketUseCase(private val basketRepository: BasketRepository, private val productId: Long) {
 
     fun invoke() {
-        basketRepository.removeSelectedProduct(productId)
+        basketRepository.getSelectedProductList().forEach {
+            if(it.productId == productId) {
+                val c: Int = it.count;
+                if(c == 1)
+                    basketRepository.removeSelectedProduct(productId)
+                else
+                    basketRepository.editSelectedProduct(SelectedProduct(productId, c - 1))
+            }
+        }
     }
 
 }
